[SITIONIX-122] - fix sonar issues in tests and repository code

+++ application/src/main/java/com/sitionix/atmssox/application/usecase/RuleSuggestionAnalysisPolicy.java
@@ -108,7 +108,7 @@ public class RuleSuggestionAnalysisPolicy {
             return true;
         }
         return lastAiSuggestionCreatedAt.get()
-                .plusSeconds((long) this.properties.getConversationCooldownMinutes() * 60L)
+                .plusSeconds(this.properties.getConversationCooldownMinutes() * 60L)
                 .isBefore(now);
     }
